+++ content_scripts/vomnibar.js
@@ -40,6 +40,13 @@ const Vomnibar = {
     this.open(sourceFrameId, options);
   },
 
+  activateTabGroupSelection(sourceFrameId) {
+    this.open(sourceFrameId, {
+      completer: "tabGroups",
+      selectFirst: true,
+    });
+  },
+
   activateBookmarksInNewTab(sourceFrameId, registryEntry) {
     const options = Object.assign({}, registryEntry.options, {
       completer: "bookmarks",
